Move coder to times.py
