+++ xarray/coding/times.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import re
+import typing
 import warnings
 from collections.abc import Callable, Hashable
 from datetime import datetime, timedelta
@@ -20,7 +21,7 @@ from xarray.coding.variables import (
     unpack_for_decoding,
     unpack_for_encoding,
 )
-from xarray.core import indexing
+from xarray.core import duck_array_ops, indexing
 from xarray.core.common import contains_cftime_datetimes, is_np_datetime_like
 from xarray.core.duck_array_ops import array_all, array_any, asarray, ravel, reshape
 from xarray.core.formatting import first_n_items, format_timestamp, last_item
@@ -1408,6 +1409,83 @@ class CFTimedeltaCoder(VariableCoder):
             )
             data = lazy_elemwise_func(data, transform, dtype=dtype)
 
+            return Variable(dims, data, attrs, encoding, fastpath=True)
+        else:
+            return variable
+
+
+class Timedelta64TypeArray(indexing.ExplicitlyIndexedNDArrayMixin):
+    """Decode arrays on the fly from integer to np.timedelta64 datatype
+
+    This is useful for decoding timedelta64 arrays from integer typed netCDF
+    variables.
+
+    >>> x = np.array([1, 0, 1, 1, 0], dtype="int64")
+
+    >>> x.dtype
+    dtype('int64')
+
+    >>> Timedelta64TypeArray(x, np.dtype("timedelta64[ns]")).dtype
+    dtype('timedelta64[ns]')
+
+    >>> indexer = indexing.BasicIndexer((slice(None),))
+    >>> Timedelta64TypeArray(x, np.dtype("timedelta64[ns]"))[indexer].dtype
+    dtype('timedelta64[ns]')
+    """
+
+    __slots__ = ("_dtype", "array")
+
+    def __init__(self, array, dtype: np.typing.DTypeLike) -> None:
+        self.array = indexing.as_indexable(array)
+        self._dtype = dtype
+
+    @property
+    def dtype(self):
+        return np.dtype(self._dtype)
+
+    def _oindex_get(self, key):
+        return np.asarray(self.array.oindex[key], dtype=self.dtype)
+
+    def _vindex_get(self, key):
+        return np.asarray(self.array.vindex[key], dtype=self.dtype)
+
+    def __getitem__(self, key) -> np.ndarray:
+        return np.asarray(self.array[key], dtype=self.dtype)
+
+
+class LiteralTimedelta64Coder(VariableCoder):
+    """Code np.timedelta64 values."""
+
+    def encode(self, variable: Variable, name: T_Name = None) -> Variable:
+        if np.issubdtype(variable.data.dtype, np.timedelta64):
+            from xarray.coding.times import _numpy_to_netcdf_timeunit
+
+            dims, data, attrs, encoding = unpack_for_encoding(variable)
+            resolution, _ = np.datetime_data(variable.dtype)
+            dtype = f"timedelta64[{resolution}]"
+            units = _numpy_to_netcdf_timeunit(resolution)
+            safe_setitem(attrs, "dtype", dtype, name=name)
+            safe_setitem(attrs, "units", units, name=name)
+            data = duck_array_ops.astype(data, dtype=np.int64, copy=True)
+            return Variable(dims, data, attrs, encoding, fastpath=True)
+        else:
+            return variable
+
+    def decode(self, variable: Variable, name: T_Name = None) -> Variable:
+        if has_timedelta64_encoding_dtype(variable.attrs):
+            dims, data, attrs, encoding = unpack_for_decoding(variable)
+            dtype = pop_to(attrs, encoding, "dtype", name=name)
+            pop_to(attrs, encoding, "units", name=name)
+            dtype = np.dtype(dtype)
+            resolution, _ = np.datetime_data(dtype)
+            if resolution not in typing.get_args(PDDatetimeUnitOptions):
+                raise ValueError(
+                    f"Following pandas, xarray only supports decoding to "
+                    f"timedelta64 values with a resolution of 's', 'ms', "
+                    f"'us', or 'ns'. Encoded values have a resolution of "
+                    f"{resolution!r}."
+                )
+            data = Timedelta64TypeArray(data, dtype)
             return Variable(dims, data, attrs, encoding, fastpath=True)
         else:
             return variable
