+++ xarray/conventions.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING, Any, Literal, TypeVar, Union
 import numpy as np
 
 from xarray.coders import CFDatetimeCoder, CFTimedeltaCoder
-from xarray.coding import strings, variables
+from xarray.coding import strings, times, variables
 from xarray.coding.variables import SerializationWarning, pop_to
 from xarray.core import indexing
 from xarray.core.common import (
@@ -98,7 +98,7 @@ def encode_cf_variable(
         variables.NonStringCoder(),
         variables.DefaultFillvalueCoder(),
         variables.BooleanCoder(),
-        variables.LiteralTimedelta64Coder(),
+        times.LiteralTimedelta64Coder(),
     ]:
         var = coder.encode(var, name=name)
 
@@ -239,7 +239,7 @@ def decode_cf_variable(
         original_dtype = var.dtype
 
     var = variables.BooleanCoder().decode(var)
-    var = variables.LiteralTimedelta64Coder().decode(var)
+    var = times.LiteralTimedelta64Coder().decode(var)
 
     dimensions, data, attributes, encoding = variables.unpack_for_decoding(var)
